+++ ci/php53.sh
@@ -1,3 +1,6 @@
+#!/usr/bin/env bash
+set -Eeuxo pipefail
+
 composer install --working-dir=$1/test/app1 --quiet || exit 1
 composer install --working-dir=$1/test/app2 --quiet || exit 1
 composer install --working-dir=$1 --quiet || exit 1